Rewise the way how we implement compression, split the client into multiple files
